SyncIntTests broken up
