+++ Darwin/Services/StitchRemoteMongoDBService/StitchRemoteMongoDBServiceTests/LocalPerformanceTestContext.swift
@@ -8,7 +8,7 @@ import StitchDarwinCoreTestUtils
 import StitchCoreLocalMongoDBService
 @testable import StitchRemoteMongoDBService
 
-class LocalPerformanceContext: SyncPerformanceContext {
+class LocalPerformanceTestContext: SyncPerformanceContext {
     let dbName: String
     let collName: String
     let userId: String
@@ -61,9 +61,13 @@ class LocalPerformanceContext: SyncPerformanceContext {
         }
 
         // swiftlint:disable force_cast
-        CoreLocalMongoDBService.shared.localInstances.forEach { client in
-            try! client.listDatabases().forEach {
-                try? client.db($0["name"] as! String).drop()
+        try CoreLocalMongoDBService.shared.localInstances.forEach { client in
+            do {
+                try client.listDatabases().forEach {
+                    try? client.db($0["name"] as! String).drop()
+                }
+            } catch {
+                throw "Could not drop databases in LocalPerformanceTestContext.teardown()"
             }
         }
         // swiftlint:enable force_cast
